double register slow condition handlers
