+++ src/sentry/workflow_engine/handlers/condition/__init__.py
@@ -3,8 +3,6 @@ __all__ = [
     "EventFrequencyCountHandler",
     "EventFrequencyPercentHandler",
     "EventSeenCountConditionHandler",
-    "EventUniqueUserFrequencyCountHandler",
-    "EventUniqueUserFrequencyPercentHandler",
     "ReappearedEventConditionHandler",
     "RegressionEventConditionHandler",
     "ExistingHighPriorityIssueConditionHandler",
@@ -28,10 +26,6 @@ from .event_attribute_handler import EventAttributeConditionHandler
 from .event_created_by_detector_handler import EventCreatedByDetectorConditionHandler
 from .event_frequency_handlers import EventFrequencyCountHandler, EventFrequencyPercentHandler
 from .event_seen_count_handler import EventSeenCountConditionHandler
-from .event_unique_user_frequency_handlers import (
-    EventUniqueUserFrequencyCountHandler,
-    EventUniqueUserFrequencyPercentHandler,
-)
 from .existing_high_priority_issue_handler import ExistingHighPriorityIssueConditionHandler
 from .first_seen_event_handler import FirstSeenEventConditionHandler
 from .issue_category_handler import IssueCategoryConditionHandler
