+++ src/sentry/workflow_engine/handlers/condition/event_frequency_handlers.py
@@ -1,19 +1,22 @@
 from collections import defaultdict
 from datetime import datetime
-from typing import Self
+from typing import Any, Self
 
 from sentry import tsdb
 from sentry.issues.constants import get_issue_tsdb_group_model
 from sentry.models.group import Group
+from sentry.rules.conditions.event_frequency import (
+    COMPARISON_INTERVALS,
+    STANDARD_INTERVALS,
+    percent_increase,
+)
 from sentry.tsdb.base import TSDBModel
 from sentry.workflow_engine.handlers.condition.event_frequency_base_handler import (
     BaseEventFrequencyConditionHandler,
-    BaseEventFrequencyCountHandler,
-    BaseEventFrequencyPercentHandler,
 )
 from sentry.workflow_engine.models.data_condition import Condition
 from sentry.workflow_engine.registry import condition_handler_registry
-from sentry.workflow_engine.types import DataConditionHandler, WorkflowJob
+from sentry.workflow_engine.types import DataConditionHandler, DataConditionResult, WorkflowJob
 
 
 class EventFrequencyConditionHandler(BaseEventFrequencyConditionHandler):
@@ -55,9 +58,49 @@ class EventFrequencyConditionHandler(BaseEventFrequencyConditionHandler):
         return batch_sums
 
 
+class BaseEventFrequencyCountHandler:
+    comparison_json_schema = {
+        "type": "object",
+        "properties": {
+            "interval": {"type": "string", "enum": list(STANDARD_INTERVALS.keys())},
+            "value": {"type": "integer", "minimum": 0},
+        },
+        "required": ["interval", "value"],
+        "additionalProperties": False,
+    }
+
+    @staticmethod
+    def evaluate_value(value: WorkflowJob, comparison: Any) -> DataConditionResult:
+        if len(value.get("snuba_results", [])) != 1:
+            return False
+        return value["snuba_results"][0] > comparison["value"]
+
+
+class BaseEventFrequencyPercentHandler:
+    comparison_json_schema = {
+        "type": "object",
+        "properties": {
+            "interval": {"type": "string", "enum": list(STANDARD_INTERVALS.keys())},
+            "value": {"type": "integer", "minimum": 0},
+            "comparison_interval": {"type": "string", "enum": list(COMPARISON_INTERVALS.keys())},
+        },
+        "required": ["interval", "value", "comparison_interval"],
+        "additionalProperties": False,
+    }
+
+    @staticmethod
+    def evaluate_value(value: WorkflowJob, comparison: Any) -> DataConditionResult:
+        if len(value.get("snuba_results", [])) != 2:
+            return False
+        return (
+            percent_increase(value["snuba_results"][0], value["snuba_results"][1])
+            > comparison["value"]
+        )
+
+
 @condition_handler_registry.register(Condition.EVENT_FREQUENCY_COUNT)
+@condition_handler_registry.register(Condition.EVENT_UNIQUE_USER_FREQUENCY_COUNT)
 class EventFrequencyCountHandler(
-    EventFrequencyConditionHandler,
     BaseEventFrequencyCountHandler,
     DataConditionHandler[WorkflowJob],
 ):
@@ -65,8 +108,8 @@ class EventFrequencyCountHandler(
 
 
 @condition_handler_registry.register(Condition.EVENT_FREQUENCY_PERCENT)
+@condition_handler_registry.register(Condition.EVENT_UNIQUE_USER_FREQUENCY_PERCENT)
 class EventFrequencyPercentHandler(
-    EventFrequencyConditionHandler,
     BaseEventFrequencyPercentHandler,
     DataConditionHandler[WorkflowJob],
 ):
